requirements.txt
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
Django==3.0.8
django-background-tasks==1.2.5
django-compat==1.0.15
django-jsonfield==1.4.0
djangorestframework==3.11.0
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.1.0
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.11
zipp==3.4.0
